Restructure packaging for econtext
COAND-497
