+++ econtext/src/main/java/com/adyen/checkout/econtext/internal/ui/model/EContextInputData.kt
@@ -6,10 +6,12 @@
  * Created by ozgur on 7/6/2022.
  */
 
-package com.adyen.checkout.econtext
+package com.adyen.checkout.econtext.internal.ui.model
 
+import androidx.annotation.RestrictTo
 import com.adyen.checkout.components.base.InputData
 
+@RestrictTo(RestrictTo.Scope.LIBRARY_GROUP)
 data class EContextInputData(
     var firstName: String = "",
     var lastName: String = "",
