+++ econtext/src/test/java/com/adyen/checkout/econtext/EContextComponentTest.kt
@@ -18,6 +18,8 @@ import com.adyen.checkout.components.PaymentComponentState
 import com.adyen.checkout.components.base.ComponentEventHandler
 import com.adyen.checkout.components.test.TestComponentViewType
 import com.adyen.checkout.core.log.Logger
+import com.adyen.checkout.econtext.internal.ui.EContextComponentViewType
+import com.adyen.checkout.econtext.internal.ui.EContextDelegate
 import com.adyen.checkout.test.TestDispatcherExtension
 import com.adyen.checkout.test.extensions.invokeOnCleared
 import kotlinx.coroutines.ExperimentalCoroutinesApi
