+++ econtext/src/test/java/com/adyen/checkout/econtext/TestEContextComponent.kt
@@ -12,6 +12,7 @@ import com.adyen.checkout.action.DefaultActionHandlingComponent
 import com.adyen.checkout.action.GenericActionDelegate
 import com.adyen.checkout.components.PaymentComponentState
 import com.adyen.checkout.components.base.ComponentEventHandler
+import com.adyen.checkout.econtext.internal.ui.EContextDelegate
 
 internal class TestEContextComponent internal constructor(
     delegate: EContextDelegate<TestEContextPaymentMethod>,
